@@ -11,7 +11,7 @@ description: Have you been turned into a problem with writing code for Http
   functionality you have been looking for.
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   build: ^2.0.0
@@ -24,6 +24,7 @@ dependencies:
   build_config: ^0.4.5
   universal_io: ^2.0.0-nullsafety.2
   dart_style: ^2.0.0
+  collection: ^1.15.0
 
 dev_dependencies:
   json_serializable: ^4.0.0
