@@ -5,8 +5,8 @@ import 'package:chopper/chopper.dart' as chopper;
 import 'example_swagger.enums.swagger.dart' as enums;
 export 'example_swagger.enums.swagger.dart';
 
-part 'example_swagger.swagger.chopper2.dart';
-part 'example_swagger.swagger.g2.dart';
+part 'example_swagger.swagger.chopper.dart';
+part 'example_swagger.swagger.g.dart';
 
 // **************************************************************************
 // SwaggerChopperGenerator
@@ -30,21 +30,21 @@ abstract class ExampleSwagger extends ChopperService {
   ///@param body Pet object that needs to be added to the store
 
   @Post(path: '/pet')
-  Future<chopper.Response> addPet({@Body() @required Pet? body});
+  Future<chopper.Response> petPost({@Body() @required Pet? body});
 
   ///Update an existing pet
   ///@param body Pet object that needs to be added to the store
 
   @Put(path: '/pet')
-  Future<chopper.Response> updatePet({@Body() @required Pet? body});
+  Future<chopper.Response> petPut({@Body() @required Pet? body});
 
   ///Finds Pets by status
   ///@param status Status values that need to be considered for filter
   ///@param color Status values that need to be considered for filter
-  Future<Response<List<Pet>>> findPetsByStatus(
+  Future<Response<List<Pet>>> petFindByStatusGet(
       {enums.PetFindByStatusGetStatus? status,
       List<enums.PetFindByStatusGetColor>? color}) {
-    return _findPetsByStatus(
+    return _petFindByStatusGet(
         status: enums.$PetFindByStatusGetStatusMap[status],
         color: color!.map((element) {
           enums.$PetFindByStatusGetColorMap[element];
@@ -52,7 +52,7 @@ abstract class ExampleSwagger extends ChopperService {
   }
 
   @Get(path: '/pet/findByStatus')
-  Future<chopper.Response<List<Pet>>> _findPetsByStatus(
+  Future<chopper.Response<List<Pet>>> _petFindByStatusGet(
       {@Query('status') @required String? status,
       @Query('color') @required List<String?>? color});
 
@@ -60,14 +60,14 @@ abstract class ExampleSwagger extends ChopperService {
   ///@param tags Tags to filter by
 
   @Get(path: '/pet/findByTags')
-  Future<chopper.Response<List<Pet>>> findPetsByTags(
+  Future<chopper.Response<List<Pet>>> petFindByTagsGet(
       {@Query('tags') @required List<String>? tags});
 
   ///Find pet by ID
   ///@param petId ID of pet to return
 
   @Get(path: '/pet/{petId}')
-  Future<chopper.Response<Pet>> getPetById(
+  Future<chopper.Response<Pet>> petPetIdGet(
       {@Path('petId') @required int? petId});
 
   ///Updates a pet in the store with form data
@@ -78,17 +78,20 @@ abstract class ExampleSwagger extends ChopperService {
 
   @Post(path: '/pet/{petId}')
   @FactoryConverter(request: FormUrlEncodedConverter.requestFactory)
-  Future<chopper.Response> updatePetWithForm(
+  Future<chopper.Response> petPetIdPost(
       {@Path('petId') @required int? petId,
       @Field('name') String? name,
       @Field('status') String? status,
       @Body() @required Object? body});
 
   ///Deletes a pet
+  ///@param api_key
   ///@param petId Pet id to delete
 
   @Delete(path: '/pet/{petId}')
-  Future<chopper.Response> deletePet({@Path('petId') @required int? petId});
+  Future<chopper.Response> petPetIdDelete(
+      {@Header('api_key') String? api_key,
+      @Path('petId') @required int? petId});
 
   ///uploads an image
   ///@param petId ID of pet to update
@@ -98,7 +101,7 @@ abstract class ExampleSwagger extends ChopperService {
 
   @Post(path: '/pet/{petId}/uploadImage')
   @FactoryConverter(request: FormUrlEncodedConverter.requestFactory)
-  Future<chopper.Response<ApiResponse>> uploadFile(
+  Future<chopper.Response<ApiResponse>> petPetIdUploadImagePost(
       {@Path('petId') @required int? petId,
       @Field('additionalMetadata') String? additionalMetadata,
       @Field('file') List<int>? file,
@@ -107,47 +110,47 @@ abstract class ExampleSwagger extends ChopperService {
   ///Returns pet inventories by status
 
   @Get(path: '/store/inventory')
-  Future<chopper.Response<List<dynamic>>> getInventory();
+  Future<chopper.Response<List<dynamic>>> storeInventoryGet();
 
   ///Place an order for a pet
   ///@param body order placed for purchasing the pet
 
   @Post(path: '/store/order')
-  Future<chopper.Response<OrderWithDash>> placeOrder(
+  Future<chopper.Response<OrderWithDash>> storeOrderPost(
       {@Body() @required Order? body});
 
   ///Find purchase order by ID
   ///@param orderId ID of pet that needs to be fetched
 
   @Get(path: '/store/order/{orderId}')
-  Future<chopper.Response<Order>> getOrderById(
+  Future<chopper.Response<Order>> storeOrderOrderIdGet(
       {@Path('orderId') @required int? orderId});
 
   ///Delete purchase order by ID
   ///@param orderId ID of the order that needs to be deleted
 
   @Delete(path: '/store/order/{orderId}')
-  Future<chopper.Response> deleteOrder(
+  Future<chopper.Response> storeOrderOrderIdDelete(
       {@Path('orderId') @required int? orderId});
 
   ///Create user
   ///@param body Created user object
 
   @Post(path: '/user')
-  Future<chopper.Response> createUser({@Body() @required User? body});
+  Future<chopper.Response> userPost({@Body() @required User? body});
 
   ///Creates list of users with given input array
   ///@param body List of user object
 
   @Post(path: '/user/createWithArray')
-  Future<chopper.Response> createUsersWithArrayInput(
+  Future<chopper.Response> userCreateWithArrayPost(
       {@Body() @required Object? body});
 
   ///Creates list of users with given input array
   ///@param body List of user object
 
   @Post(path: '/user/createWithList')
-  Future<chopper.Response> createUsersWithListInput(
+  Future<chopper.Response> userCreateWithListPost(
       {@Body() @required Object? body});
 
   ///Logs user into the system
@@ -155,20 +158,20 @@ abstract class ExampleSwagger extends ChopperService {
   ///@param password The password for login in clear text
 
   @Get(path: '/user/login')
-  Future<chopper.Response<String>> loginUser(
+  Future<chopper.Response<String>> userLoginGet(
       {@Query('username') @required String? username,
       @Query('password') @required String? password});
 
   ///Logs out current logged in user session
 
   @Get(path: '/user/logout')
-  Future<chopper.Response> logoutUser();
+  Future<chopper.Response> userLogoutGet();
 
   ///Get user by user name
   ///@param username The name that needs to be fetched. Use user1 for testing.
 
   @Get(path: '/user/{username}')
-  Future<chopper.Response<User>> getUserByName(
+  Future<chopper.Response<User>> userUsernameGet(
       {@Path('username') @required String? username});
 
   ///Updated user
@@ -176,7 +179,7 @@ abstract class ExampleSwagger extends ChopperService {
   ///@param body Updated user object
 
   @Put(path: '/user/{username}')
-  Future<chopper.Response> updateUser(
+  Future<chopper.Response> userUsernamePut(
       {@Path('username') @required String? username,
       @Body() @required User? body});
 
@@ -184,7 +187,7 @@ abstract class ExampleSwagger extends ChopperService {
   ///@param username The name that needs to be deleted
 
   @Delete(path: '/user/{username}')
-  Future<chopper.Response> deleteUser(
+  Future<chopper.Response> userUsernameDelete(
       {@Path('username') @required String? username});
 }
 
@@ -202,13 +205,13 @@ final Map<Type, Object Function(Map<String, dynamic>)>
 @JsonSerializable(explicitToJson: true)
 class Order {
   Order({
-    this.id,
-    this.petId,
-    this.quantity,
-    this.shipDateTime,
-    this.shipDate,
-    this.status,
-    this.complete,
+    required this.id,
+    required this.petId,
+    required this.quantity,
+    required this.shipDateTime,
+    required this.shipDate,
+    required this.status,
+    required this.complete,
   });
 
   factory Order.fromJson(Map<String, dynamic> json) => _$OrderFromJson(json);
@@ -259,12 +262,12 @@ extension $OrderExtension on Order {
 @JsonSerializable(explicitToJson: true)
 class OrderWithDash {
   OrderWithDash({
-    this.id,
-    this.petId,
-    this.quantity,
-    this.shipDate,
-    this.status,
-    this.complete,
+    required this.id,
+    required this.petId,
+    required this.quantity,
+    required this.shipDate,
+    required this.status,
+    required this.complete,
   });
 
   factory OrderWithDash.fromJson(Map<String, dynamic> json) =>
@@ -312,8 +315,8 @@ extension $OrderWithDashExtension on OrderWithDash {
 @JsonSerializable(explicitToJson: true)
 class Category {
   Category({
-    this.id,
-    this.name,
+    required this.id,
+    required this.name,
   });
 
   factory Category.fromJson(Map<String, dynamic> json) =>
@@ -321,7 +324,7 @@ class Category {
 
   @JsonKey(name: 'id', includeIfNull: false, defaultValue: 36)
   final int? id;
-  @JsonKey(name: 'name', includeIfNull: false)
+  @JsonKey(name: 'name', includeIfNull: false, defaultValue: '')
   final String? name;
   static const fromJsonFactory = _$CategoryFromJson;
   static const toJsonFactory = _$CategoryToJson;
@@ -337,31 +340,31 @@ extension $CategoryExtension on Category {
 @JsonSerializable(explicitToJson: true)
 class User {
   User({
-    this.id,
-    this.username,
-    this.firstName,
-    this.lastName,
-    this.email,
-    this.password,
-    this.phone,
-    this.userStatus,
+    required this.id,
+    required this.username,
+    required this.firstName,
+    required this.lastName,
+    required this.email,
+    required this.password,
+    required this.phone,
+    required this.userStatus,
   });
 
   factory User.fromJson(Map<String, dynamic> json) => _$UserFromJson(json);
 
   @JsonKey(name: 'id', includeIfNull: false, defaultValue: 36)
   final int? id;
-  @JsonKey(name: 'username', includeIfNull: false)
+  @JsonKey(name: 'username', includeIfNull: false, defaultValue: '')
   final String? username;
-  @JsonKey(name: 'firstName', includeIfNull: false)
+  @JsonKey(name: 'firstName', includeIfNull: false, defaultValue: '')
   final String? firstName;
-  @JsonKey(name: 'lastName', includeIfNull: false)
+  @JsonKey(name: 'lastName', includeIfNull: false, defaultValue: '')
   final String? lastName;
-  @JsonKey(name: 'email', includeIfNull: false)
+  @JsonKey(name: 'email', includeIfNull: false, defaultValue: '')
   final String? email;
-  @JsonKey(name: 'password', includeIfNull: false)
+  @JsonKey(name: 'password', includeIfNull: false, defaultValue: '')
   final String? password;
-  @JsonKey(name: 'phone', includeIfNull: false)
+  @JsonKey(name: 'phone', includeIfNull: false, defaultValue: '')
   final String? phone;
   @JsonKey(name: 'userStatus', includeIfNull: false, defaultValue: 36)
   final int? userStatus;
@@ -395,15 +398,15 @@ extension $UserExtension on User {
 @JsonSerializable(explicitToJson: true)
 class Tag {
   Tag({
-    this.id,
-    this.name,
+    required this.id,
+    required this.name,
   });
 
   factory Tag.fromJson(Map<String, dynamic> json) => _$TagFromJson(json);
 
   @JsonKey(name: 'id', includeIfNull: false, defaultValue: 36)
   final int? id;
-  @JsonKey(name: 'name', includeIfNull: false)
+  @JsonKey(name: 'name', includeIfNull: false, defaultValue: '')
   final String? name;
   static const fromJsonFactory = _$TagFromJson;
   static const toJsonFactory = _$TagToJson;
@@ -419,12 +422,12 @@ extension $TagExtension on Tag {
 @JsonSerializable(explicitToJson: true)
 class Pet {
   Pet({
-    this.id,
+    required this.id,
     this.category,
-    this.name,
-    this.photoUrls,
-    this.tags,
-    this.status,
+    required this.name,
+    required this.photoUrls,
+    required this.tags,
+    required this.status,
   });
 
   factory Pet.fromJson(Map<String, dynamic> json) => _$PetFromJson(json);
@@ -433,12 +436,12 @@ class Pet {
   final int? id;
   @JsonKey(name: 'category', includeIfNull: false)
   final Category? category;
-  @JsonKey(name: 'name', includeIfNull: false)
+  @JsonKey(name: 'name', includeIfNull: false, defaultValue: '')
   final String? name;
-  @JsonKey(name: 'photoUrls', includeIfNull: false, defaultValue: <String>[])
-  final List<String?>? photoUrls;
+  @JsonKey(name: 'photoUrls', includeIfNull: false, defaultValue: <dynamic>[])
+  final List<dynamic>? photoUrls;
   @JsonKey(name: 'tags', includeIfNull: false, defaultValue: <Tag>[])
-  final List<Tag?>? tags;
+  final List<Tag>? tags;
   @JsonKey(
       name: 'status',
       includeIfNull: false,
@@ -455,8 +458,8 @@ extension $PetExtension on Pet {
       {int? id,
       Category? category,
       String? name,
-      List<String?>? photoUrls,
-      List<Tag?>? tags,
+      List<dynamic>? photoUrls,
+      List<Tag>? tags,
       enums.PetStatus? status}) {
     return Pet(
         id: id ?? this.id,
@@ -471,9 +474,9 @@ extension $PetExtension on Pet {
 @JsonSerializable(explicitToJson: true)
 class ApiResponse {
   ApiResponse({
-    this.code,
-    this.type,
-    this.message,
+    required this.code,
+    required this.type,
+    required this.message,
   });
 
   factory ApiResponse.fromJson(Map<String, dynamic> json) =>
@@ -481,9 +484,9 @@ class ApiResponse {
 
   @JsonKey(name: 'code', includeIfNull: false, defaultValue: 36)
   final int? code;
-  @JsonKey(name: 'type', includeIfNull: false)
+  @JsonKey(name: 'type', includeIfNull: false, defaultValue: '')
   final String? type;
-  @JsonKey(name: 'message', includeIfNull: false)
+  @JsonKey(name: 'message', includeIfNull: false, defaultValue: '')
   final String? message;
   static const fromJsonFactory = _$ApiResponseFromJson;
   static const toJsonFactory = _$ApiResponseToJson;
@@ -510,28 +513,25 @@ enums.PetFindByStatusGetStatus petFindByStatusGetStatusFromJson(
     return enums.PetFindByStatusGetStatus.swaggerGeneratedUnknown;
   }
 
-  if (!enums.$PetFindByStatusGetStatusMap.entries
-      .contains(petFindByStatusGetStatus)) {
-    return enums.PetFindByStatusGetStatus.swaggerGeneratedUnknown;
-  }
-
   return enums.$PetFindByStatusGetStatusMap.entries
-      .firstWhere((element) => element.value == petFindByStatusGetStatus)
+      .firstWhere((element) => element.value == petFindByStatusGetStatus,
+          orElse: () => const MapEntry(
+              enums.PetFindByStatusGetStatus.swaggerGeneratedUnknown, ''))
       .key;
 }
 
-List<String?> petFindByStatusGetStatusListToJson(
+List<String> petFindByStatusGetStatusListToJson(
     List<enums.PetFindByStatusGetStatus>? petFindByStatusGetStatus) {
   if (petFindByStatusGetStatus == null) {
     return [];
   }
 
   return petFindByStatusGetStatus
-      .map((e) => enums.$PetFindByStatusGetStatusMap[e])
+      .map((e) => enums.$PetFindByStatusGetStatusMap[e]!)
       .toList();
 }
 
-List<enums.PetFindByStatusGetStatus?> petFindByStatusGetStatusListFromJson(
+List<enums.PetFindByStatusGetStatus> petFindByStatusGetStatusListFromJson(
     List? petFindByStatusGetStatus) {
   if (petFindByStatusGetStatus == null) {
     return [];
@@ -553,28 +553,25 @@ enums.PetFindByStatusGetColor petFindByStatusGetColorFromJson(
     return enums.PetFindByStatusGetColor.swaggerGeneratedUnknown;
   }
 
-  if (!enums.$PetFindByStatusGetColorMap.entries
-      .contains(petFindByStatusGetColor)) {
-    return enums.PetFindByStatusGetColor.swaggerGeneratedUnknown;
-  }
-
   return enums.$PetFindByStatusGetColorMap.entries
-      .firstWhere((element) => element.value == petFindByStatusGetColor)
+      .firstWhere((element) => element.value == petFindByStatusGetColor,
+          orElse: () => const MapEntry(
+              enums.PetFindByStatusGetColor.swaggerGeneratedUnknown, ''))
       .key;
 }
 
-List<String?> petFindByStatusGetColorListToJson(
+List<String> petFindByStatusGetColorListToJson(
     List<enums.PetFindByStatusGetColor>? petFindByStatusGetColor) {
   if (petFindByStatusGetColor == null) {
     return [];
   }
 
   return petFindByStatusGetColor
-      .map((e) => enums.$PetFindByStatusGetColorMap[e])
+      .map((e) => enums.$PetFindByStatusGetColorMap[e]!)
       .toList();
 }
 
-List<enums.PetFindByStatusGetColor?> petFindByStatusGetColorListFromJson(
+List<enums.PetFindByStatusGetColor> petFindByStatusGetColorListFromJson(
     List? petFindByStatusGetColor) {
   if (petFindByStatusGetColor == null) {
     return [];
@@ -594,24 +591,22 @@ enums.OrderStatus orderStatusFromJson(String? orderStatus) {
     return enums.OrderStatus.swaggerGeneratedUnknown;
   }
 
-  if (!enums.$OrderStatusMap.entries.contains(orderStatus)) {
-    return enums.OrderStatus.swaggerGeneratedUnknown;
-  }
-
   return enums.$OrderStatusMap.entries
-      .firstWhere((element) => element.value == orderStatus)
+      .firstWhere((element) => element.value == orderStatus,
+          orElse: () =>
+              const MapEntry(enums.OrderStatus.swaggerGeneratedUnknown, ''))
       .key;
 }
 
-List<String?> orderStatusListToJson(List<enums.OrderStatus>? orderStatus) {
+List<String> orderStatusListToJson(List<enums.OrderStatus>? orderStatus) {
   if (orderStatus == null) {
     return [];
   }
 
-  return orderStatus.map((e) => enums.$OrderStatusMap[e]).toList();
+  return orderStatus.map((e) => enums.$OrderStatusMap[e]!).toList();
 }
 
-List<enums.OrderStatus?> orderStatusListFromJson(List? orderStatus) {
+List<enums.OrderStatus> orderStatusListFromJson(List? orderStatus) {
   if (orderStatus == null) {
     return [];
   }
@@ -630,27 +625,25 @@ enums.OrderWithDashStatus orderWithDashStatusFromJson(
     return enums.OrderWithDashStatus.swaggerGeneratedUnknown;
   }
 
-  if (!enums.$OrderWithDashStatusMap.entries.contains(orderWithDashStatus)) {
-    return enums.OrderWithDashStatus.swaggerGeneratedUnknown;
-  }
-
   return enums.$OrderWithDashStatusMap.entries
-      .firstWhere((element) => element.value == orderWithDashStatus)
+      .firstWhere((element) => element.value == orderWithDashStatus,
+          orElse: () => const MapEntry(
+              enums.OrderWithDashStatus.swaggerGeneratedUnknown, ''))
       .key;
 }
 
-List<String?> orderWithDashStatusListToJson(
+List<String> orderWithDashStatusListToJson(
     List<enums.OrderWithDashStatus>? orderWithDashStatus) {
   if (orderWithDashStatus == null) {
     return [];
   }
 
   return orderWithDashStatus
-      .map((e) => enums.$OrderWithDashStatusMap[e])
+      .map((e) => enums.$OrderWithDashStatusMap[e]!)
       .toList();
 }
 
-List<enums.OrderWithDashStatus?> orderWithDashStatusListFromJson(
+List<enums.OrderWithDashStatus> orderWithDashStatusListFromJson(
     List? orderWithDashStatus) {
   if (orderWithDashStatus == null) {
     return [];
@@ -670,24 +663,22 @@ enums.PetStatus petStatusFromJson(String? petStatus) {
     return enums.PetStatus.swaggerGeneratedUnknown;
   }
 
-  if (!enums.$PetStatusMap.entries.contains(petStatus)) {
-    return enums.PetStatus.swaggerGeneratedUnknown;
-  }
-
   return enums.$PetStatusMap.entries
-      .firstWhere((element) => element.value == petStatus)
+      .firstWhere((element) => element.value == petStatus,
+          orElse: () =>
+              const MapEntry(enums.PetStatus.swaggerGeneratedUnknown, ''))
       .key;
 }
 
-List<String?> petStatusListToJson(List<enums.PetStatus>? petStatus) {
+List<String> petStatusListToJson(List<enums.PetStatus>? petStatus) {
   if (petStatus == null) {
     return [];
   }
 
-  return petStatus.map((e) => enums.$PetStatusMap[e]).toList();
+  return petStatus.map((e) => enums.$PetStatusMap[e]!).toList();
 }
 
-List<enums.PetStatus?> petStatusListFromJson(List? petStatus) {
+List<enums.PetStatus> petStatusListFromJson(List? petStatus) {
   if (petStatus == null) {
     return [];
   }
@@ -724,7 +715,7 @@ class CustomJsonDecoder {
       return throw "Could not find factory for type $T. Is '$T: $T.fromJsonFactory' included in the CustomJsonDecoder instance creation in bootstrapper.dart?";
     }
 
-    return jsonFactory(values) as T;
+    return jsonFactory(values);
   }
 
   List<T> _decodeList<T>(Iterable values) =>
@@ -749,6 +740,7 @@ class JsonSerializableConverter extends chopper.JsonConverter {
 
 final jsonDecoder = CustomJsonDecoder(ExampleSwaggerJsonDecoderMappings);
 
+// ignore: unused_element
 String? _dateToJson(DateTime? date) {
   if (date == null) {
     return null;
