@@ -8,52 +8,52 @@ part 'swagger_request_parameter.g2.dart';
 @JsonSerializable()
 class SwaggerRequestParameter {
   SwaggerRequestParameter(
-      {this.inParameter,
-      this.name,
-      this.description,
-      this.isRequired,
-      this.type,
+      {this.inParameter ='',
+      this.name = '',
+      this.description = '',
+      this.isRequired = false,
+      this.type = '',
       this.item,
-      this.collectionFormat,
+      this.collectionFormat = '',
       this.items,
       this.schema,
-      this.ref,
-      this.key});
+      this.ref = '',
+      this.key = ''});
 
   factory SwaggerRequestParameter.fromJson(Map<String, dynamic> json) =>
       _$SwaggerRequestParameterFromJson(json);
 
-  @JsonKey(name: 'in')
+  @JsonKey(name: 'in', defaultValue: '')
   String inParameter;
 
-  @JsonKey(name: 'name')
+  @JsonKey(name: 'name', defaultValue: '')
   String name;
 
-  @JsonKey(name: 'description')
+  @JsonKey(name: 'description', defaultValue: '')
   String description;
 
   @JsonKey(name: 'required', defaultValue: false)
   bool isRequired;
 
-  @JsonKey(name: '\$ref')
+  @JsonKey(name: '\$ref', defaultValue: '')
   String ref;
 
-  @JsonKey(name: 'type')
+  @JsonKey(name: 'type', defaultValue: '')
   String type;
 
   @JsonKey(name: 'item')
-  ParameterItem item;
+  ParameterItem? item;
 
-  @JsonKey(name: 'collectionFormat')
+  @JsonKey(name: 'collectionFormat', defaultValue: '')
   String collectionFormat;
 
   @JsonKey(name: 'schema')
-  SwaggerParameterSchema schema;
+  SwaggerParameterSchema? schema;
 
   @JsonKey(name: 'items')
-  SwaggerRequestItems items;
+  SwaggerRequestItems? items;
 
-  @JsonKey(ignore: true)
+  @JsonKey(ignore: true, defaultValue: '')
   String key;
 
   Map<String, dynamic> toJson() => _$SwaggerRequestParameterToJson(this);
