@@ -13,17 +13,18 @@ void main() {
     const fileName = 'order_service.dart';
 
     test('Should parse object name as a field Type', () {
-      final result = generator.generate(
-                    model_with_parameters_v3, fileName, GeneratorOptions());
+      final result = generator.generate(model_with_parameters_v3, fileName,
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
-
-      expect(
-          result, contains('final enums.TokensResponseTokenType tokenType'));
+      expect(result, contains('final enums.TokensResponseTokenType? tokenType'));
     });
 
     test('Should generate .toLower() when caseSensitive: false', () {
-      final result = generator.generate(model_with_parameters_v3, fileName,
-          GeneratorOptions(enumsCaseSensitive: false));
+      final result = generator.generate(
+          model_with_parameters_v3,
+          fileName,
+          GeneratorOptions(
+              enumsCaseSensitive: false, inputFolder: '', outputFolder: ''));
 
       expect(
           result,
@@ -32,31 +33,37 @@ void main() {
     });
 
     test('Should NOT generate .toLower() when caseSensitive: false', () {
-      final result = generator.generate(model_with_parameters_v3, fileName,
-          GeneratorOptions(enumsCaseSensitive: true));
+      final result = generator.generate(
+          model_with_parameters_v3,
+          fileName,
+          GeneratorOptions(
+              enumsCaseSensitive: true, inputFolder: '', outputFolder: ''));
 
       expect(result, contains('element.value == someEnumModel'));
     });
 
     test('Should parse object name as a field Type', () {
-      final result = generator2.generate(
-          model_with_parameters_v2, fileName, GeneratorOptions());
+      final result = generator2.generate(model_with_parameters_v2, fileName,
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(
           result,
           contains(
-              'final enums.ActiveOrderAndListSummaryShoppingType shoppingType'));
+              'final enums.ActiveOrderAndListSummaryShoppingType? shoppingType'));
     });
 
     test('Should parse object name as a field Type', () {
       const expectedResult = "@JsonKey(name: 'expires_in', defaultValue: 19)";
-      final generatorOptions =
-          GeneratorOptions(defaultValuesMap: <DefaultValueMap>[
-        DefaultValueMap(
-          typeName: 'int',
-          defaultValue: '19',
-        )
-      ]);
+      final generatorOptions = GeneratorOptions(
+        defaultValuesMap: <DefaultValueMap>[
+          DefaultValueMap(
+            typeName: 'int',
+            defaultValue: '19',
+          )
+        ],
+        inputFolder: '',
+        outputFolder: '',
+      );
       final result = generator.generate(
           model_with_parameters_v3, fileName, generatorOptions);
 
@@ -125,7 +132,7 @@ void main() {
       const parameterName = 'orderId';
       const expectedResult = 'Object';
       final result =
-          generator.getParameterTypeName(className, parameterName, null);
+          generator.getParameterTypeName(className, parameterName, {});
 
       expect(result, contains(expectedResult));
     });
@@ -147,7 +154,7 @@ void main() {
       const refNameParameter = 'animals_Object';
       const expectedResult = 'AnimalsObject';
       final result = generator.getParameterTypeName(
-          className, parameterName, null, refNameParameter);
+          className, parameterName, {}, refNameParameter);
 
       expect(result, contains(expectedResult));
     });
@@ -207,7 +214,11 @@ void main() {
       const propertyName = 'shipDate';
       const jsonKeyExpendedResult = "@JsonKey(name: '$propertyName'";
       final result = generator.generatePropertyContentByDefault(
-          propertyEntryMap, propertyName, [], [], GeneratorOptions());
+          propertyEntryMap,
+          propertyName,
+          [],
+          [],
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, contains(jsonKeyExpendedResult));
     });
@@ -221,7 +232,9 @@ void main() {
           [],
           [],
           GeneratorOptions(
-              includeIfNull: IncludeIfNull(enabled: true, value: false)));
+              includeIfNull: IncludeIfNull(enabled: true, value: false),
+              inputFolder: '',
+              outputFolder: ''));
 
       expect(result, contains(', includeIfNull: false'));
     });
@@ -235,7 +248,9 @@ void main() {
           [],
           [],
           GeneratorOptions(
-              includeIfNull: IncludeIfNull(enabled: false, value: false)));
+              includeIfNull: IncludeIfNull(enabled: false, value: false),
+              inputFolder: '',
+              outputFolder: ''));
 
       expect(result.contains(', includeIfNull: false'), equals(false));
     });
@@ -258,7 +273,7 @@ void main() {
         className,
         [],
         [],
-        GeneratorOptions(),
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         {},
       );
 
@@ -279,7 +294,7 @@ void main() {
         className,
         ['enums.Pet'],
         [],
-        GeneratorOptions(),
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         {},
       );
 
@@ -290,14 +305,13 @@ void main() {
   group('generateEnumPropertyContent', () {
     test('Should return enum property with JsonKey with fromJson and toJson',
         () {
-      // ignore: prefer_const_declarations
-      final key = 'dog';
+      const key = 'dog';
       const className = 'animals';
       const jsonKeyExpectedResult =
           'toJson: animalsDogToJson, fromJson: animalsDogFromJson';
-      const expectedResult = 'final AnimalsDog dog;';
-      final result = generator.generateEnumPropertyContent(
-          key, className, [], [], GeneratorOptions());
+      const expectedResult = 'final AnimalsDog? dog;';
+      final result = generator.generateEnumPropertyContent(key, className, [],
+          [], GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, contains(jsonKeyExpectedResult));
       expect(result, contains(expectedResult));
@@ -315,12 +329,12 @@ void main() {
       final result = generator.generateModelClassContent(
           className,
           map,
-          null,
+          {},
           <DefaultValueMap>[],
           useDefaultNullForLists,
           [],
           [],
-          GeneratorOptions());
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, contains(classExpectedResult));
       expect(result, contains(factoryConstructorExpectedResult));
@@ -336,12 +350,12 @@ void main() {
       final result = generator2.generateModelClassContent(
           className,
           map,
-          null,
+          {},
           <DefaultValueMap>[],
           useDefaultNullForLists,
           [],
           [],
-          GeneratorOptions());
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, contains(classExpectedResult));
       expect(result, contains(factoryConstructorExpectedResult));
@@ -353,8 +367,8 @@ void main() {
       const expectedResult = '';
       final result = generator.generateConstructorPropertiesContent(
         '',
-        null,
-        GeneratorOptions(),
+        {},
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         [],
         [],
         [],
@@ -370,7 +384,12 @@ void main() {
       const expectedResult = 'this.animal';
       final result = generator.generateConstructorPropertiesContent(
         '',
-          map, GeneratorOptions(), [], [], []);
+        map,
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
+        [],
+        [],
+        [],
+      );
 
       expect(result, contains(expectedResult));
     });
@@ -393,7 +412,7 @@ void main() {
         className,
         [],
         [],
-        GeneratorOptions(),
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         {},
       );
 
@@ -416,7 +435,7 @@ void main() {
         className,
         ['enums.Pet'],
         [],
-        GeneratorOptions(),
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         {},
       );
 
@@ -432,8 +451,15 @@ void main() {
       const className = 'Animals';
       const jsonKeyExpectedResult = "\t@JsonKey(name: 'Animals')\n";
       const fieldExpectedResult = 'final Pet? animals';
-      final result = generator.generatePropertiesContent(map, null, className,
-          <DefaultValueMap>[], false, [], [], GeneratorOptions());
+      final result = generator.generatePropertiesContent(
+          map,
+          {},
+          className,
+          <DefaultValueMap>[],
+          false,
+          [],
+          [],
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, contains(jsonKeyExpectedResult));
       expect(result, contains(fieldExpectedResult));
@@ -449,8 +475,15 @@ void main() {
       const className = 'Animals';
       const jsonKeyExpectedResult = "\t@JsonKey(name: 'Animals')\n";
       const fieldExpectedResult = 'final Pet? animals';
-      final result = generator.generatePropertiesContent(map, null, className,
-          <DefaultValueMap>[], false, [], [], GeneratorOptions());
+      final result = generator.generatePropertiesContent(
+          map,
+          {},
+          className,
+          <DefaultValueMap>[],
+          false,
+          [],
+          [],
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, contains(jsonKeyExpectedResult));
       expect(result, contains(fieldExpectedResult));
@@ -464,8 +497,15 @@ void main() {
       const className = 'Animals';
       const jsonKeyExpectedResult = "\t@JsonKey(name: 'Animals')\n";
       const fieldExpectedResult = 'final Pet? animals';
-      final result = generator.generatePropertiesContent(map, null, className,
-          <DefaultValueMap>[], false, [], [], GeneratorOptions());
+      final result = generator.generatePropertiesContent(
+          map,
+          {},
+          className,
+          <DefaultValueMap>[],
+          false,
+          [],
+          [],
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, contains(jsonKeyExpectedResult));
       expect(result, contains(fieldExpectedResult));
@@ -479,8 +519,15 @@ void main() {
       const className = 'Animals';
       const jsonKeyExpectedResult = "\t@JsonKey(name: '\$with')\n";
       const fieldExpectedResult = 'final Pet? \$with';
-      final result = generator.generatePropertiesContent(map, null, className,
-          <DefaultValueMap>[], false, [], [], GeneratorOptions());
+      final result = generator.generatePropertiesContent(
+          map,
+          {},
+          className,
+          <DefaultValueMap>[],
+          false,
+          [],
+          [],
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, contains(jsonKeyExpectedResult));
       expect(result, contains(fieldExpectedResult));
@@ -496,7 +543,7 @@ void main() {
       const jsonKeyExpectedResult =
           "@JsonKey(name: 'Dog', defaultValue: <Object>[])";
 
-      const propertyExpectedResult = 'final List<Object> dog';
+      const propertyExpectedResult = 'final List<Object>? dog';
       final result = generator.generateListPropertyContent(
         propertyName,
         propertyKey,
@@ -505,7 +552,7 @@ void main() {
         false,
         [],
         [],
-        GeneratorOptions(),
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         {},
       );
 
@@ -528,7 +575,7 @@ void main() {
         false,
         ['Dog'],
         [],
-        GeneratorOptions(),
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         {},
       );
 
@@ -552,11 +599,11 @@ void main() {
         false,
         [],
         [],
-        GeneratorOptions(),
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         {},
       );
 
-      expect(result, contains('final List<TestOriginalRef> dog;'));
+      expect(result, contains('final List<TestOriginalRef>? dog;'));
     });
   });
 
@@ -575,7 +622,7 @@ void main() {
         false,
         [],
         [],
-        GeneratorOptions(),
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         {},
       );
 
@@ -598,26 +645,26 @@ void main() {
         false,
         [],
         [],
-        GeneratorOptions(),
+        GeneratorOptions(inputFolder: '', outputFolder: ''),
         {},
       );
 
-      expect(result, contains('final List<Object> dog;'));
+      expect(result, contains('final List<Object>? dog;'));
     });
   });
 
   group('getModelInheritedProperties', () {
     test('Should generate 2 levels of inheritance', () {
-      final result = generator.generate(
-          model_with_inheritance, 'MyClass', GeneratorOptions());
+      final result = generator.generate(model_with_inheritance, 'MyClass',
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(
           result, contains('class ExtendedErrorModel extends BasicErrorModel'));
     });
 
     test('Should generate 3 levels of inheritance', () {
-      final result = generator.generate(
-          model_with_inheritance_3_levels, 'MyClass', GeneratorOptions());
+      final result = generator.generate(model_with_inheritance_3_levels,
+          'MyClass', GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result,
           contains('class MostExtendedErrorModel extends ExtendedErrorModel'));
@@ -645,23 +692,27 @@ void main() {
           {'\$ref': 'ClassNameName'},
           ['enums.ClassNameName'],
           [],
-          GeneratorOptions());
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
-      expect(result, contains('final enums.ClassNameName name;'));
+      expect(result, contains('final enums.ClassNameName? name;'));
     });
   });
 
   group('Tests for generateResponses', () {
     test('Should generate empty string for V2', () {
       final result = generator2.generateResponses(
-          schemas_responses_with_response, 'fileName', GeneratorOptions());
+          schemas_responses_with_response,
+          'fileName',
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, equals(''));
     });
 
     test('Should generate class from responses V3', () {
       final result = generator.generateResponses(
-          schemas_responses_with_response, 'fileName', GeneratorOptions());
+          schemas_responses_with_response,
+          'fileName',
+          GeneratorOptions(inputFolder: '', outputFolder: ''));
 
       expect(result, contains('class SpaResponse'));
     });
@@ -670,7 +721,10 @@ void main() {
       final result = generator.generateResponses(
           schemas_responses_with_response_and_schemas,
           'fileName',
-          GeneratorOptions());
+          GeneratorOptions(
+            inputFolder: '',
+            outputFolder: '',
+          ));
 
       expect(result, contains('class SpaResponse'));
     });
