@@ -7,6 +7,7 @@ import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_
 import 'package:swagger_dart_code_generator/src/swagger_models/requests/swagger_request_parameter.dart';
 import 'package:swagger_dart_code_generator/src/swagger_models/swagger_path.dart';
 import 'package:swagger_dart_code_generator/src/swagger_models/swagger_root.dart';
+import 'package:collection/collection.dart';
 
 abstract class SwaggerEnumsGenerator {
   static const String defaultEnumFieldName = 'value_';
@@ -23,7 +24,7 @@ abstract class SwaggerEnumsGenerator {
 
     final enumsFromRequestBodies = generateEnumsFromRequestBodies(requestBodies);
 
-    if (definitions == null) {
+    if (definitions.isEmpty) {
       return '$enumsFromRequests$enumsFromResponses$enumsFromRequestBodies';
     }
 
@@ -57,15 +58,15 @@ $enumsFromRequestBodies
   }
 
   String generateEnumsFromResponses(Map<String, dynamic> responses) {
-    if (responses == null) {
+    if (responses.isEmpty) {
       return '';
     }
 
     final enumsFromResponses = responses.keys
         .map((String className) {
           final response = responses[className];
-          final content = response['content'] as Map<String, dynamic>;
-          final firstContent = content?.entries?.first?.value;
+          final content = response['content'] as Map<String, dynamic>?;
+          final firstContent = content != null && content.entries.isNotEmpty ? content.entries.first.value : null;
           final schema = firstContent == null ? null : firstContent['schema'];
 
           if (schema == null) {
@@ -84,15 +85,15 @@ $enumsFromRequestBodies
   }
 
   String generateEnumsFromRequestBodies(Map<String, dynamic> requestBodies) {
-    if (requestBodies == null) {
+    if (requestBodies.isEmpty) {
       return '';
     }
 
     final enumsFromRequestBodies = requestBodies.keys
         .map((String className) {
           final response = requestBodies[className];
-          final content = response['content'] as Map<String, dynamic>;
-          final firstContent = content?.entries?.first?.value;
+          final content = response['content'] as Map<String, dynamic>?;
+          final firstContent = content != null && content.entries.isNotEmpty ? content.entries.first.value : null;
           final schema = firstContent == null ? null : firstContent['schema'];
 
           if (schema == null) {
@@ -113,7 +114,7 @@ $enumsFromRequestBodies
   static SwaggerRequestParameter getOriginalOrOverriddenRequestParameter(
       SwaggerRequestParameter swaggerRequestParameter,
       List<SwaggerRequestParameter> definedParameters) {
-    if (swaggerRequestParameter.ref == null || definedParameters == null) {
+    if (swaggerRequestParameter.ref.isEmpty|| definedParameters.isEmpty) {
       return swaggerRequestParameter;
     }
 
@@ -140,7 +141,7 @@ $enumsFromRequestBodies
         swaggerRequest.parameters = swaggerRequest.parameters
             .map((SwaggerRequestParameter parameter) =>
                 getOriginalOrOverriddenRequestParameter(
-                    parameter, swaggerRoot.components?.parameters))
+                    parameter, swaggerRoot.components?.parameters ?? []))
             .toList();
       });
     });
@@ -151,7 +152,7 @@ $enumsFromRequestBodies
       for (var j = 0; j < swaggerPath.requests.length; j++) {
         final swaggerRequest = swaggerPath.requests[j];
 
-        if (swaggerRequest.parameters == null) {
+        if (swaggerRequest.parameters.isEmpty) {
           continue;
         }
 
@@ -170,9 +171,9 @@ $enumsFromRequestBodies
           }
 
           final enumValues = swaggerRequestParameter.schema?.enumValues ??
-              swaggerRequestParameter.items?.enumValues;
+              swaggerRequestParameter.items?.enumValues ?? [];
 
-          if (enumValues != null) {
+          if (enumValues.isNotEmpty) {
             final enumContent = generateEnumContent(name, enumValues);
 
             result.writeln(enumContent);
@@ -221,7 +222,7 @@ $enumMap
     neededValues.addAll(enumValues);
 
     final unknownEnumPart =
-        ',\n\t$enumName.swaggerGeneratedUnknown: \'swaggerGeneratedUnknown\'';
+        ',\n\t$enumName.swaggerGeneratedUnknown: \'\'';
 
     final result = neededValues
             .map((String enumFieldName) =>
@@ -233,7 +234,10 @@ $enumMap
   }
 
   String getValidatedEnumFieldName(String name) {
-    name ??= 'null';
+    if(name.isEmpty)
+    {
+      name = 'null';
+    }
 
     var result = name
         .replaceAll(RegExp(r'[^\w|\_|)]'), '_')
@@ -264,7 +268,7 @@ $enumMap
         swaggerRequest.parameters = swaggerRequest.parameters
             .map((SwaggerRequestParameter parameter) =>
                 getOriginalOrOverriddenRequestParameter(
-                    parameter, swaggerRoot.components?.parameters))
+                    parameter, swaggerRoot.components?.parameters ?? []))
             .toList();
       });
     });
@@ -275,7 +279,7 @@ $enumMap
       for (var j = 0; j < swaggerPath.requests.length; j++) {
         final swaggerRequest = swaggerPath.requests[j];
 
-        if (swaggerRequest.parameters == null) {
+        if (swaggerRequest.parameters.isEmpty) {
           continue;
         }
 
@@ -292,9 +296,9 @@ $enumMap
           }
 
           final enumValues = swaggerRequestParameter.schema?.enumValues ??
-              swaggerRequestParameter.items?.enumValues;
+              swaggerRequestParameter.items?.enumValues ?? [];
 
-          if (enumValues != null) {
+          if (enumValues.isNotEmpty) {
             enumNames.add(name);
           }
         }
@@ -306,7 +310,7 @@ $enumMap
 
   String generateEnumsContentFromModelProperties(
       Map<String, dynamic> map, String className) {
-    if (map == null) {
+    if (map.isEmpty) {
       return '';
     }
 
@@ -363,7 +367,7 @@ $enumMap
   String generateEnumValuesContent(List<dynamic> values) {
     return values
         .map((dynamic e) =>
-            "\t@JsonValue('${e.toString().replaceAll("\$", "\\\$")}')\n  ${getValidatedEnumFieldName(e?.toString())}")
+            "\t@JsonValue('${e.toString().replaceAll("\$", "\\\$")}')\n  ${getValidatedEnumFieldName(e.toString())}")
         .join(',\n');
   }
 
@@ -383,20 +387,20 @@ $enumMap
 
     if (map.containsKey('allOf')) {
       final allOf = map['allOf'] as List<dynamic>;
-      var propertiesContainer = allOf.firstWhere(
+      var propertiesContainer = allOf.firstWhereOrNull(
           (e) => (e as Map<String, dynamic>).containsKey('properties'),
-          orElse: () => null) as Map<String, dynamic>;
+         ) as Map<String, dynamic>? ?? {};
 
-      if (propertiesContainer != null) {
-        properties = propertiesContainer['properties'] as Map<String, dynamic>;
+      if (propertiesContainer.isNotEmpty) {
+        properties = propertiesContainer['properties'] as Map<String, dynamic>? ?? {};
       } else {
-        properties = map['properties'] as Map<String, dynamic>;
+        properties = map['properties'] as Map<String, dynamic>? ?? {};
       }
     } else {
-      properties = map['properties'] as Map<String, dynamic>;
+      properties = map['properties'] as Map<String, dynamic>? ?? {};
     }
 
-    if (properties == null) {
+    if (properties.isEmpty) {
       return '';
     }
 
