@@ -20,6 +20,7 @@ abstract class SwaggerModelsGenerator {
   String generate(String dartCode, String fileName, GeneratorOptions options);
   String generateResponses(
       String dartCode, String fileName, GeneratorOptions options);
+
   String generateRequestBodies(
       String dartCode, String fileName, GeneratorOptions options);
   Map<String, dynamic> getModelProperties(Map<String, dynamic> modelMap);
@@ -80,7 +81,7 @@ abstract class SwaggerModelsGenerator {
             allEnumsNames, options.enumsCaseSensitive)
         : '';
 
-    if (classes == null) {
+    if (classes.isEmpty) {
       return '';
     }
 
@@ -112,7 +113,7 @@ abstract class SwaggerModelsGenerator {
   }
 
   static String getValidatedClassName(String className) {
-    if (className == null) {
+    if (className.isEmpty) {
       return className;
     }
 
@@ -140,12 +141,12 @@ abstract class SwaggerModelsGenerator {
 
   String getParameterTypeName(
       String className, String parameterName, Map<String, dynamic> parameter,
-      [String refNameParameter]) {
+      [String? refNameParameter]) {
     if (refNameParameter != null) {
       return refNameParameter.pascalCase;
     }
 
-    if (parameter == null) {
+    if (parameter.isEmpty) {
       return 'Object';
     }
 
@@ -153,7 +154,7 @@ abstract class SwaggerModelsGenerator {
       return parameter['\$ref'].toString().split('/').last.pascalCase;
     }
 
-    switch (parameter['type'] as String) {
+    switch (parameter['type'] as String?) {
       case 'integer':
       case 'int':
         return 'int';
@@ -172,7 +173,7 @@ abstract class SwaggerModelsGenerator {
       case 'object':
         return 'Object';
       case 'array':
-        final items = parameter['items'] as Map<String, dynamic>;
+        final items = parameter['items'] as Map<String, dynamic>? ?? {};
         return getParameterTypeName(className, parameterName, items);
       default:
         if (parameter['oneOf'] != null) {
@@ -242,11 +243,11 @@ abstract class SwaggerModelsGenerator {
   }
 
   String generateIncludeIfNullString(GeneratorOptions options) {
-    if (options.includeIfNull == null || !options.includeIfNull.enabled) {
+    if (options.includeIfNull == null || !options.includeIfNull!.enabled) {
       return '';
     }
 
-    return ', includeIfNull: ${options.includeIfNull.value}';
+    return ', includeIfNull: ${options.includeIfNull!.value}';
   }
 
   String generatePropertyContentByDefault(
@@ -255,9 +256,12 @@ abstract class SwaggerModelsGenerator {
       List<String> allEnumNames,
       List<String> allEnumListNames,
       GeneratorOptions options) {
-    final typeName =
-        getValidatedClassName(propertyEntryMap['originalRef'].toString()) ??
-            'dynamic';
+    var typeName =
+        getValidatedClassName(propertyEntryMap['originalRef'].toString());
+
+    if (typeName.isEmpty) {
+      typeName = 'dynamic';
+    }
 
     final unknownEnumValue = generateUnknownEnumValue(
         allEnumNames, allEnumListNames, typeName.toString(), false);
@@ -293,8 +297,8 @@ abstract class SwaggerModelsGenerator {
   }
 
   String generateToJsonForDate(Map<String, dynamic> map) {
-    final type = map['type']?.toString()?.toLowerCase();
-    final format = map['format']?.toString()?.toLowerCase();
+    final type = map['type']?.toString().toLowerCase();
+    final format = map['format']?.toString().toLowerCase();
 
     final isDate = type == 'string' && format == 'date';
 
@@ -319,7 +323,7 @@ abstract class SwaggerModelsGenerator {
 
     String typeName;
     if (basicTypesMap.containsKey(parameterName)) {
-      typeName = basicTypesMap[parameterName];
+      typeName = basicTypesMap[parameterName]!;
     } else {
       typeName = getValidatedClassName(getParameterTypeName(
           className, propertyName, propertyEntryMap, parameterName));
@@ -342,10 +346,6 @@ abstract class SwaggerModelsGenerator {
     final jsonKeyContent =
         "@JsonKey(name: '$propertyKey'$includeIfNullString$unknownEnumValue$dateToJsonValue)\n";
 
-    if (typeName.startsWith('enums.')) {
-      return '\t$jsonKeyContent\tfinal $typeName ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
-    }
-
     return '\t$jsonKeyContent\tfinal $typeName? ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
   }
 
@@ -361,7 +361,7 @@ abstract class SwaggerModelsGenerator {
     final parameterName = propertyEntryMap['\$ref'].toString().split('/').last;
     String typeName;
     if (basicTypesMap.containsKey(parameterName)) {
-      typeName = basicTypesMap[parameterName];
+      typeName = basicTypesMap[parameterName]!;
     } else {
       typeName = getValidatedClassName(getParameterTypeName(
           className, propertyName, propertyEntryMap, parameterName));
@@ -386,12 +386,7 @@ abstract class SwaggerModelsGenerator {
     final jsonKeyContent =
         "@JsonKey(name: '$propertyKey'$includeIfNullString$unknownEnumValue)\n";
 
-    if ((!typeName.startsWith('List<') || options.useDefaultNullForLists) &&
-        !typeName.startsWith('enums.')) {
-      return '\t$jsonKeyContent\tfinal $typeName? ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
-    }
-
-    return '\t$jsonKeyContent\tfinal $typeName ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
+    return '\t$jsonKeyContent\tfinal $typeName? ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
   }
 
   String generateEnumPropertyContent(
@@ -407,13 +402,13 @@ abstract class SwaggerModelsGenerator {
     allEnumNames.add(enumName);
 
     final unknownEnumValue = generateUnknownEnumValue(
-        allEnumNames, allEnumListNames, enumName, false);
+        allEnumNames, allEnumListNames, enumName, false).substring(2);
 
     final includeIfNullString = generateIncludeIfNullString(options);
 
     return '''
   @JsonKey($unknownEnumValue$includeIfNullString)
-  final ${className.capitalize + key.capitalize} ${SwaggerModelsGenerator.generateFieldName(key)};''';
+  final ${className.capitalize + key.capitalize}? ${SwaggerModelsGenerator.generateFieldName(key)};''';
   }
 
   String generateListPropertyContent(
@@ -426,20 +421,20 @@ abstract class SwaggerModelsGenerator {
       List<String> allEnumListNames,
       GeneratorOptions options,
       Map<String, String> basicTypesMap) {
-    final dynamic items = propertyEntryMap['items'];
+    final items = propertyEntryMap['items'];
 
-    String typeName;
+    var typeName = '';
     if (items != null) {
-      typeName = getValidatedClassName(items['originalRef'] as String);
+      typeName = getValidatedClassName(items['originalRef'] as String? ?? '');
 
-      if (typeName == null) {
-        final ref = items['\$ref'] as String;
-        typeName = ref?.split('/')?.last;
+      if (typeName.isEmpty) {
+        final ref = items['\$ref'] as String?;
+        typeName = ref?.split('/').last ?? 'dynamic';
 
         if (basicTypesMap.containsKey(typeName)) {
-          typeName = basicTypesMap[typeName];
-        } else {
-          typeName = typeName?.pascalCase;
+          typeName = basicTypesMap[typeName]!;
+        } else if(typeName != 'dynamic') {
+          typeName = typeName.pascalCase;
         }
       }
 
@@ -448,8 +443,10 @@ abstract class SwaggerModelsGenerator {
       }
     }
 
-    typeName ??= getParameterTypeName(
-        className, propertyName, items as Map<String, dynamic>);
+    if (typeName.isEmpty) {
+      typeName = getParameterTypeName(
+          className, propertyName, items as Map<String, dynamic>? ?? {});
+    }
 
     final unknownEnumValue = generateUnknownEnumValue(
         allEnumNames, allEnumListNames, typeName, true);
@@ -465,11 +462,7 @@ abstract class SwaggerModelsGenerator {
           "@JsonKey(name: '$propertyKey'$includeIfNullString$unknownEnumValue)\n";
     }
 
-    if (options.useDefaultNullForLists && !typeName.startsWith('enums.')) {
-      return '''  $jsonKeyContent  final List<$typeName?>? ${SwaggerModelsGenerator.generateFieldName(propertyName)};''';
-    }
-
-    return '''  $jsonKeyContent  final List<$typeName> ${SwaggerModelsGenerator.generateFieldName(propertyName)};''';
+    return '$jsonKeyContent  final List<$typeName>? ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
   }
 
   String generateGeneralPropertyContent(
@@ -502,27 +495,20 @@ abstract class SwaggerModelsGenerator {
     jsonKeyContent += unknownEnumValue;
     jsonKeyContent += dateToJsonValue;
 
-    var hasDefaultValue = false;
-
     if ((val['type'] == 'bool' || val['type'] == 'boolean') &&
         val['default'] != null) {
       jsonKeyContent += ', defaultValue: ${val['default']})\n';
-      hasDefaultValue = true;
     } else if (defaultValues
         .any((DefaultValueMap element) => element.typeName == typeName)) {
       final defaultValue = defaultValues.firstWhere(
           (DefaultValueMap element) => element.typeName == typeName);
       jsonKeyContent +=
           ', defaultValue: ${generateDefaultValueFromMap(defaultValue)})\n';
-      hasDefaultValue = true;
     } else {
       jsonKeyContent += ')\n';
     }
 
-    final shouldBeNotNullable =
-        hasDefaultValue || typeName.startsWith('enums.');
-
-    return '''  $jsonKeyContent  final $typeName${shouldBeNotNullable ? '' : '?'} ${SwaggerModelsGenerator.generateFieldName(propertyName)};''';
+    return '''  $jsonKeyContent  final $typeName? ${SwaggerModelsGenerator.generateFieldName(propertyName)};''';
   }
 
   String generatePropertyContentByType(
@@ -549,7 +535,6 @@ abstract class SwaggerModelsGenerator {
           options,
           basicTypesMap,
         );
-        break;
       case 'enum':
         return generateEnumPropertyContent(
           propertyName,
@@ -558,7 +543,6 @@ abstract class SwaggerModelsGenerator {
           allEnumListNames,
           options,
         );
-        break;
       default:
         return generateGeneralPropertyContent(
           propertyName,
@@ -582,7 +566,7 @@ abstract class SwaggerModelsGenerator {
       List<String> allEnumNames,
       List<String> allEnumListNames,
       GeneratorOptions options) {
-    if (propertiesMap == null) {
+    if (propertiesMap.isEmpty) {
       return '';
     }
 
@@ -646,7 +630,7 @@ abstract class SwaggerModelsGenerator {
   static Map<String, String> generateBasicTypesMapFromSchemas(
       Map<String, dynamic> schemas) {
     final result = <String, String>{};
-    if (schemas == null) {
+    if (schemas.isEmpty) {
       return result;
     }
 
@@ -682,7 +666,7 @@ abstract class SwaggerModelsGenerator {
       case 'boolean':
         return 'bool';
       default:
-        return null;
+        return '';
     }
   }
 
@@ -711,12 +695,9 @@ enums.$neededName ${neededName.camelCase}FromJson(String? ${neededName.camelCase
     return enums.$neededName.swaggerGeneratedUnknown;
   }
 
-  if (!enums.\$${neededName}Map.entries.contains(${neededName.camelCase})) {
-    return enums.$neededName.swaggerGeneratedUnknown;
-  }
-
   return enums.\$${neededName}Map.entries
-      .firstWhere((element) => element.value$toLowerCaseString == ${neededName.camelCase}$toLowerCaseString)
+      .firstWhere((element) => element.value$toLowerCaseString == ${neededName.camelCase}$toLowerCaseString,
+      orElse: () => const MapEntry(enums.$neededName.swaggerGeneratedUnknown, ''))
       .key;
 }
 
@@ -756,7 +737,7 @@ List<enums.$neededName> ${neededName.camelCase}ListFromJson(
     List<String> allEnumNames,
     List<String> allEnumListNames,
   ) {
-    if (entityMap == null) {
+    if (entityMap.isEmpty) {
       return '';
     }
 
@@ -768,12 +749,13 @@ List<enums.$neededName> ${neededName.camelCase}ListFromJson(
       final hasDefaultValue = value['default'] != null ||
           defaultValues.any((element) =>
               element.typeName ==
-              _mapBasicTypeToDartType(value['type'].toString(), null));
+              _mapBasicTypeToDartType(value['type'].toString(), ''));
 
       final isList = value['type'] == 'array' ||
           allEnumListNames.contains('enums.${key.pascalCase}');
 
-      final type = value['\$ref']?.toString()?.split('/')?.last?.pascalCase ?? key;
+      final type =
+          value['\$ref']?.toString().split('/').last.pascalCase ?? key;
 
       final isEnum = allEnumNames.contains('enums.${type.pascalCase}') ||
           allEnumNames.contains('enums.${className + type.pascalCase}');
