@@ -1,6 +1,6 @@
 extension CapitalizeExtension on String {
   String get capitalize {
-    return this == null || isEmpty
+    return isEmpty
         ? this
         : (this[0].toUpperCase() + substring(1));
   }
