@@ -5,25 +5,25 @@ part 'swagger_parameter_schema.g2.dart';
 @JsonSerializable()
 class SwaggerParameterSchema {
   SwaggerParameterSchema(
-      {this.type, this.enumValues, this.originalRef, this.ref});
+      {this.type = '', this.enumValues = const [], this.originalRef = '', this.ref = ''});
 
   factory SwaggerParameterSchema.fromJson(Map<String, dynamic> json) =>
       _$SwaggerParameterSchemaFromJson(json);
 
-  @JsonKey(name: 'type')
+  @JsonKey(name: 'type', defaultValue: '')
   String type;
 
   @JsonKey(name: 'enum', fromJson: enumsFromJson)
   List<String> enumValues;
 
-  @JsonKey(name: 'originalRef')
+  @JsonKey(name: 'originalRef', defaultValue: '')
   String originalRef;
 
-  @JsonKey(name: '\$ref')
+  @JsonKey(name: '\$ref', defaultValue: '')
   String ref;
 
   Map<String, dynamic> toJson() => _$SwaggerParameterSchemaToJson(this);
 
-  static List<String> enumsFromJson(List list) =>
-      list?.map((e) => e.toString())?.toList();
+  static List<String> enumsFromJson(List? list) =>
+      list?.map((e) => e.toString()).toList() ?? [];
 }
