@@ -8,8 +8,11 @@ part of 'swagger_request_items.dart';
 
 SwaggerRequestItems _$SwaggerRequestItemsFromJson(Map<String, dynamic> json) {
   return SwaggerRequestItems(
-    type: json['type'] as String,
-  )..enumValues = (json['enum'] as List)?.map((e) => e as String)?.toList();
+    type: json['type'] as String? ?? '',
+    enumValues:
+        (json['enum'] as List<dynamic>?)?.map((e) => e as String).toList() ??
+            [],
+  );
 }
 
 Map<String, dynamic> _$SwaggerRequestItemsToJson(
