@@ -9,15 +9,15 @@ part of 'swagger_request_parameter.dart';
 SwaggerRequestParameter _$SwaggerRequestParameterFromJson(
     Map<String, dynamic> json) {
   return SwaggerRequestParameter(
-    inParameter: json['in'] as String,
-    name: json['name'] as String,
-    description: json['description'] as String,
-    isRequired: json['required'] as bool ?? false,
-    type: json['type'] as String,
+    inParameter: json['in'] as String? ?? '',
+    name: json['name'] as String? ?? '',
+    description: json['description'] as String? ?? '',
+    isRequired: json['required'] as bool? ?? false,
+    type: json['type'] as String? ?? '',
     item: json['item'] == null
         ? null
         : ParameterItem.fromJson(json['item'] as Map<String, dynamic>),
-    collectionFormat: json['collectionFormat'] as String,
+    collectionFormat: json['collectionFormat'] as String? ?? '',
     items: json['items'] == null
         ? null
         : SwaggerRequestItems.fromJson(json['items'] as Map<String, dynamic>),
@@ -25,7 +25,8 @@ SwaggerRequestParameter _$SwaggerRequestParameterFromJson(
         ? null
         : SwaggerParameterSchema.fromJson(
             json['schema'] as Map<String, dynamic>),
-  )..ref = json['\$ref'] as String;
+    ref: json[r'$ref'] as String? ?? '',
+  );
 }
 
 Map<String, dynamic> _$SwaggerRequestParameterToJson(
@@ -35,7 +36,7 @@ Map<String, dynamic> _$SwaggerRequestParameterToJson(
       'name': instance.name,
       'description': instance.description,
       'required': instance.isRequired,
-      '\$ref': instance.ref,
+      r'$ref': instance.ref,
       'type': instance.type,
       'item': instance.item,
       'collectionFormat': instance.collectionFormat,
