@@ -5,21 +5,21 @@ part 'swagger_info.g2.dart';
 @JsonSerializable()
 class SwaggerInfo {
   SwaggerInfo(
-      {this.description, this.version, this.title, this.termsOfService});
+      {this.description = '', this.version = '', this.title = '', this.termsOfService = ''});
 
   factory SwaggerInfo.fromJson(Map<String, dynamic> json) =>
       _$SwaggerInfoFromJson(json);
 
-  @JsonKey(name: 'description')
+  @JsonKey(name: 'description', defaultValue: '')
   String description;
 
-  @JsonKey(name: 'version')
+  @JsonKey(name: 'version', defaultValue: '')
   String version;
 
-  @JsonKey(name: 'title')
+  @JsonKey(name: 'title', defaultValue: '')
   String title;
 
-  @JsonKey(name: 'termsOfService')
+  @JsonKey(name: 'termsOfService', defaultValue: '')
   String termsOfService;
 
   Map<String, dynamic> toJson() => _$SwaggerInfoToJson(this);
