@@ -8,10 +8,10 @@ part of 'swagger_info.dart';
 
 SwaggerInfo _$SwaggerInfoFromJson(Map<String, dynamic> json) {
   return SwaggerInfo(
-    description: json['description'] as String,
-    version: json['version'] as String,
-    title: json['title'] as String,
-    termsOfService: json['termsOfService'] as String,
+    description: json['description'] as String? ?? '',
+    version: json['version'] as String? ?? '',
+    title: json['title'] as String? ?? '',
+    termsOfService: json['termsOfService'] as String? ?? '',
   );
 }
 
